Invalidate iterator on transaction clear
